requirements.txt
aiohttp-cors==0.8.1
aiortc==1.10.1
apturl==0.5.2
asyncio==3.4.3
autogluon==1.3.1
bcrypt==3.2.0
brlapi==0.8.3
catboost==1.2.8
chronos==0.3
chronos-forecasting==1.5.2
cliapp==1.20180812.1
cmdtest==0.32+git
colorful==0.5.6
command-not-found==0.3
crochet==2.1.1
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro-info==1.1+ubuntu0.2
djangorestframework==3.15.2
python-dotenv==1.0.0
schedule==1.2.0
joblib==1.3.2
duplicity==0.8.21
einops==0.8.1
fastai==2.8.2
fasteners==0.14.1
firebase-admin==6.5.0
flask-cors==5.0.1
flask-login==0.6.3
flask-sqlalchemy==3.1.1
google-generativeai==0.8.5
gradio==4.19.1
grpcio-status==1.62.2
hyperopt==0.2.7
imutils==0.5.4
ipython==8.37.0
jaraco.collections==5.1.0
jeepney==0.7.1
language-selector==0.1
launchpadlib==1.10.16
lightgbm==4.6.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
monotonic==1.6
moviepy==1.0.3
mysql-connector-python==9.3.0
nbformat==5.10.4
netifaces==0.11.0
neuralforecast==3.0.1
neuralprophet==0.9.0
nixtla==0.6.6
oauthlib==3.2.0
olefile==0.46
openai==1.78.0
opencensus==0.11.4
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
pip-chill==1.0.3
polars==1.33.1
prometheus-client==0.22.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py-spy==0.4.0
pycups==2.0.1
pygobject==3.42.1
pyinstaller==6.14.1
pymacaroons==0.13.0
pymongo==4.15.1
pynacl==1.5.0
pyodbc==5.2.0
pyrfc3339==1.1
pysftp==0.2.9
pysocks==1.7.1
puttykeys
python-apt==2.4.0+ubuntu3
python-debian==0.1.43+ubuntu1.1
pyxdg==0.27
qrcode==7.4.2
reportlab==3.6.8
scrapy==2.13.1
secretstorage==3.3.1
selenium==4.27.1
sentence-transformers==5.1.0
sentencepiece==0.2.0
shellingham==1.5.4
sqlite-vec==0.1.6
ssh-import-id==5.11
streamlit==1.31.1
sumy==0.11.0
systemd-python==234
tensorboardx==2.6.2.2
timegpt==0.0.1
tk==0.1.0
ttystatus==0.38
u8darts==0.35.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.3.96
unattended-upgrades==0.1
usb-creator==0.3.7
virtualenv==20.31.2
wadllib==1.3.6
xdg==5
xkit==0.0.0
yfinance==0.2.63
zipp==1.0.0
zope.interface==7.2
